requirements.txt
pandas
python-dotenv
imapclient
beautifulsoup4
lxml